[REF] reorganize the code to split the commands from main loop
