+++ scripts/commands/__init__.py
@@ -1,0 +1,3 @@
+from .list_pr import list_pr
+from .push import push
+from .update import update
