+++ scripts/const.py
@@ -2,3 +2,4 @@ from os import path
 
 USER_HOME = path.expanduser('~')
 CONFIG_FILE_PATH = path.join(USER_HOME, ".spConfig.ini")
+DIFF_VALID_PATH = ["src"]
